OCPBUGS-59156: Add sharding to SNO upgrade & serial presubmits.
Recent trends of the AWS upgrade & serial presubmits for SNO show a consistent runtime deep
into the 3 hour mark. This has resulted in a large amount of the recent runs being canceled
as they trip over the 4 hour time limit. This change attempts to alleviate that pressure,
using the sharding mechanism introduced in
https://docs.google.com/document/d/14uXl_JF-n7WvJj7xQOmn7wRm71LM7Fbuw-2QnP_-bTs/edit.